% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multimainfunction}
\alias{multimainfunction}
\title{Extracting factors for conditional log-linear model}
\usage{
multimainfunction(pars, newcombined)
}
\arguments{
\item{pars}{a set of parameters}

\item{newcombined}{a \code{[(2+nc) x m ]} matrix comprised of outcomes (first row), treatments (second row), and confounders (from the third row), where \code{nc} is the number of confounders.}
}
\value{
a sum of factors.
}
\description{
This is an auxiliary function to print out the factors for conditional log-linear model
given edge information.
}
