% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipartition}
\alias{multipartition}
\title{Calculating normalizing constant in conditional log-linear model.}
\usage{
multipartition(pars, combined, permutetab)
}
\arguments{
\item{pars}{a set of parameters}

\item{combined}{a \code{[(2+nc) x m ]} matrix comprised of outcomes (first row), treatments (second row), and confounders (from the third row), where \code{nc} is the number of confounders.}

\item{permutetab}{a matrix comprised of every possible values for outcome in each row.}
}
\value{
a normalizing constant
}
\description{
Calculating normalizing constant in conditional log-linear model.
}
