% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCG.R
\name{plotCG}
\alias{plotCG}
\title{Plot chain graph components from DAG}
\usage{
plotCG(DAG, vertex_attrs = list(), edge_attrs = list())
}
\arguments{
\item{DAG}{A DAG object from \code{simcausal} R Package.}

\item{vertex_attrs}{A list of parameters for DAG vertices which are passed on to \code{add.vertices} in \code{igraph}.}

\item{edge_attrs}{A list of parameters for DAG edges which are passed on to \code{add.edges} in \code{igraph}.}
}
\value{
plot
}
\description{
Plot chain graph components from DAG
}
\examples{
library(netchain)
library(simcausal)
D <- DAG.empty()
D <- D +
node("A1", t = 0, distr = "rbern", prob = 0.3) + 
node("A2", t = 0, distr = "rbern", prob = 0.5) + 
node("A3", t = 0, distr = "rbern", prob = 0.7) +
node("Y1", t = 0, distr = "rbern", prob = plogis(-1 + A1[0])) + 
node("Y2", t = 0, distr = "rbern", prob = plogis(0 - 0.5*A2[0])) + 
node("Y3", t = 0, distr = "rbern", prob = plogis(1 - 0.5*A3[0])) + 
node("Y4", t = 0, distr = "rbern", prob = plogis(1 - 0.5*A3[0]))
t.end <- 20
D <- D + 
node("Y1", t = 1:t.end, distr = "rbern",
prob = plogis(3*Y1[t-1] - 1.5 + A1[0] + Y2[t-1])) + 
node("Y2", t = 1:t.end, distr = "rbern",
prob = plogis(3*Y2[t-1] - 2 - 0.5*A1[0] + Y1[t-1])) +
node("Y3", t = 1:t.end, distr = "rbern", 
prob = plogis(3*Y3[t-1] - 1 - 0.5*A2[0] -Y4[t-1])) + 
node("Y4", t = 1:t.end, distr = "rbern", 
prob = plogis(3*Y4[t-1] - 1 - 0.5*A3[0] -Y3[t-1]))
lDAG <- set.DAG(D)
plotCG(lDAG, vertex_attrs = list(label.color = "hotpink"))


}
