% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multimainfunction}
\alias{multimainfunction}
\title{Extracting factors for conditional log-linear model}
\usage{
multimainfunction(pars, newcombined, edgeY, edgeAY, edgeExtra)
}
\arguments{
\item{pars}{a set of parameters}

\item{newcombined}{a \code{[(2+nc) x m ]} matrix comprised of outcomes (first row), treatments (second row), and confounders (from the third row), where \code{nc} is the number of confounders.}

\item{edgeY}{a matrix of which each row indicates a pair of index for adjacent outcomes.}

\item{edgeAY}{a matrix of which each row indicates a index for treatment (first column) and for outcome (second column) on which the treatment has a direct effect.}

\item{edgeExtra}{a list of edges of which a list of matrix specifying additional directed edges (from confounders or treatment to the outcomes) information.}
}
\value{
a sum of factors.
}
\description{
This is an auxiliary function to print out the factors for conditional log-linear model
given edge information.
}
